[WIP] Major refactor. Remove globus from module names
