+++ aero/jobs/user_flow.py
@@ -4,7 +4,7 @@ from typing import TypeAlias
 
 from globus_sdk import SpecificFlowClient
 
-from aero.globus.globus_auth import get_authorizer
+from aero.globus.auth import get_authorizer
 from aero.globus.utils import FLOW_IDS
 from aero.globus.utils import FlowEnum
 from aero.globus.utils import _flow_scopes
